mcp-tools/web/requirements.txt
mcp[cli]>=1.0.0
httpx>=0.28.0
beautifulsoup4>=4.12.0
html2text>=2024.2.26
# Vertex AI search grounding — ADC via the GCE VM's service account
# (metadata server), no API key.  ``google-auth`` drives auth,
# ``requests`` is the sync transport used by its ``Request`` adapter.
google-auth>=2.30.0
requests>=2.31.0
